test: split tests
